+++ src/Smartstore.Modules/Smartstore.PayPal/Providers/PayPalGooglePayProvider.cs
@@ -13,7 +13,7 @@ namespace Smartstore.PayPal.Providers
     [SystemName(PayPalConstants.GooglePay)]
     [FriendlyName("PayPal Google Pay")]
     [Order(1)]
-    [PaymentMethodType(PaymentMethodType = PaymentMethodType.Standard | PaymentMethodType.Button)]
+    [PaymentMethod(PaymentMethodType.Standard | PaymentMethodType.Button)]
     public class PayPalGooglePayProvider : PayPalProviderBase
     {
         public PayPalGooglePayProvider(